refactor: party mode をコードベースから全廃する
