+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "conductor"
-version = "0.100.0"
+version = "0.101.0"
 edition = "2024"
 
 [dependencies]
