+++ src/app/mod.rs
@@ -200,12 +200,6 @@ pub struct App {
     /// Alt+/ で出す、各パネル上の番号バッジ (2 秒で自動的に消える)。
     pub panel_number_overlay: PanelNumberOverlay,
 
-    // パーティーモード (隠しイースターエッグ)
-    /// trueのとき、UIは全開のパーティー状態になる: フォーカス中パネルの境界が
-    /// 流れる虹色に光り、シンタックストークンが虹色になり、タイトルバーが
-    /// きらめき、紙吹雪が画面を漂う。コマンドパレットから切り替える。
-    /// 永続化はしない（セッション限定の秘密）。
-    pub party_mode: bool,
 
     // リフロー・トランスクリプトビュー
     /// 無限スクロールバックモード中にClaude PTYパネルへオーバーレイされる、
